Java 1.5 migration: removed unnecessary unboxing.

+++ source/com/mucommander/ant/macosx/DateValue.java
@@ -148,26 +148,26 @@ public class DateValue implements InfoElement {
                     day = now.get(Calendar.DAY_OF_MONTH);
             }
         }
-        else if(year.intValue() < 0)
+        else if(year < 0)
             throw new BuildException("Illegal year value: " + year);
 
         // Makes sure month is initialised properly.
         if(month != null) {
-            if(month.intValue() < 1 || month.intValue() > 12)
+            if(month < 1 || month > 12)
                 throw new BuildException("Illegal month value: " + month);
             // Makes sure day is initialised properly.
-            if(day != null && (day.intValue() < 0 || !checkDay(year.intValue(), month.intValue(), day.intValue())))
+            if(day != null && (day < 0 || !checkDay(year, month, day)))
                 throw new BuildException("Illegal day value: " + day);
         }
 
         // Makes sure time is initialised properly.
         if(hours != null) {
-            if(hours.intValue() < 0 || hours.intValue() > 24)
+            if(hours < 0 || hours > 24)
                 throw new BuildException("Illegal hours value: " + hours);
             if(minutes != null) {
-                if(minutes.intValue() < 0 || minutes.intValue() > 60)
+                if(minutes < 0 || minutes > 60)
                     throw new BuildException("Illegal minutes value: " + minutes);
-                if(seconds != null && (seconds.intValue() < 0 || seconds.intValue() > 60))
+                if(seconds != null && (seconds < 0 || seconds > 60))
                     throw new BuildException("Illegal seconds value: " + seconds);
             }
         }
