+++ source/com/mucommander/ui/dialog/file/ProgressDialog.java
@@ -536,7 +536,7 @@ public class ProgressDialog extends FocusDialog implements ActionListener, ItemL
                 // Calculate the maximum bytes per second of all the samples to be displayed
                 long maxBps = 0;
                 for(int i=firstSample; i<firstSample+nbLines; i++) {
-                    long sample = ((Long)samples.elementAt(i)).longValue();
+                    long sample = (Long) samples.elementAt(i);
                     if(sample>maxBps)
                         maxBps = sample;
                 }
@@ -558,7 +558,7 @@ public class ProgressDialog extends FocusDialog implements ActionListener, ItemL
                 Polygon p = new Polygon();
                 int sampleOffset = firstSample;
                 for(int l=0; l<nbLines; l++) {
-                    p.addPoint(x, height-STROKE_WIDTH-(int)(((Long)samples.elementAt(sampleOffset++)).longValue()/yRatio));
+                    p.addPoint(x, height-STROKE_WIDTH-(int)((Long) samples.elementAt(sampleOffset++) /yRatio));
                     x+=LINE_SPACING;
                 }
                 p.addPoint(x-LINE_SPACING, height-1);
@@ -570,8 +570,8 @@ public class ProgressDialog extends FocusDialog implements ActionListener, ItemL
                 x = STROKE_WIDTH;
                 sampleOffset = firstSample;
                 for(int l=0; l<nbLines-1; l++) {
-                    g.drawLine(x, height-STROKE_WIDTH-(int)(((Long)samples.elementAt(sampleOffset)).longValue()/yRatio),
-                              (x+=LINE_SPACING), height-STROKE_WIDTH-(int)(((Long)samples.elementAt(++sampleOffset)).longValue()/yRatio));
+                    g.drawLine(x, height-STROKE_WIDTH-(int)((Long) samples.elementAt(sampleOffset) /yRatio),
+                              (x+=LINE_SPACING), height-STROKE_WIDTH-(int)((Long) samples.elementAt(++sampleOffset) /yRatio));
                 }
 
                 // Draw an horizontal line at the bottom of the graph
