+++ source/com/mucommander/ui/dialog/startup/CheckVersionDialog.java
@@ -176,7 +176,7 @@ public class CheckVersionDialog extends QuestionDialog implements Runnable {
         int actions[] = new int[nbChoices];
         String labels[] = new String[nbChoices];
         for(int i=0; i<nbChoices; i++) {
-            actions[i] = ((Integer)actionsV.elementAt(i)).intValue();
+            actions[i] = (Integer) actionsV.elementAt(i);
             labels[i] = (String)labelsV.elementAt(i);
         }
